requirements.txt
streamlit
torch
torchvision
fpdf
pillow
altair
